Update layer aplication, viewmodels
