+++ eShopSolution.ViewModels/Catalog/Products/GetManagerProductPagingRequest.cs
@@ -1,11 +1,9 @@
 ﻿using eShopSolution.ViewModels.Common;
-using System;
 using System.Collections.Generic;
-using System.Text;
 
-namespace eShopSolution.ViewModels.Catalog.Products.Manage
+namespace eShopSolution.ViewModels.Catalog.Products
 {
-    public class GetProductPagingRequest : PagingRequestBase
+    public class GetManagerProductPagingRequest : PagingRequestBase
     {
         public string Keyword { get; set; }
         public List<int> CategoryIds { get; set; }
