+++ eShopSolution.ViewModels/Catalog/Products/ProductCreateRequest.cs
@@ -1,13 +1,12 @@
 ﻿using Microsoft.AspNetCore.Http;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace eShopSolution.ViewModels.Catalog.Products.Manage
+namespace eShopSolution.ViewModels.Catalog.Products
 {
-    public class ProductUpdateRequest
+    public class ProductCreateRequest
     {
-        public int Id { get; set; }
+        public decimal Price { get; set; }
+        public decimal OriginalPrice { get; set; }
+        public int Stock { get; set; }
         public string Name { set; get; }
         public string Description { set; get; }
         public string Details { set; get; }
@@ -15,6 +14,7 @@ namespace eShopSolution.ViewModels.Catalog.Products.Manage
         public string SeoTitle { set; get; }
         public string SeoAlias { get; set; }
         public string LanguageId { set; get; }
+
         public IFormFile ThumbnailImage { get; set; }
     }
 }
